Remove photo repos
