+++ Core/Repositories/IProductRepository.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Warehouse.Core.Models;
 
@@ -6,6 +7,7 @@ namespace Warehouse.Core.Repositories
     public interface IProductRepository
     {
         Task<Product> GetProduct(int id, bool includeRelated = true);
+        Task<IEnumerable<ProductPhoto>> GetPhotos();
         Task Add(Product product);
     }
 }
