+++ Core/Repositories/ITechnicianRepository.cs
@@ -7,6 +7,7 @@ namespace Warehouse.Core.Repositories
     public interface ITechnicianRepository
     {
         Task<IEnumerable<Technician>> GetTechnicians();
+        Task<IEnumerable<TechnicianPhoto>> GetPhoto(int id);
         Task<Technician> GetTechnician(int id);
         Task Add(Technician technician);
     }
