+++ Persistence/Repositories/TechnicianRepository.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Warehouse.Core.Models;
@@ -20,11 +21,18 @@ namespace Warehouse.Persistence.Repositories
             return await _context.Technicians.ToListAsync();
         }
 
+        public async Task<IEnumerable<TechnicianPhoto>> GetPhoto(int technicianId)
+        {
+            return await _context.TechnicianPhotos
+                .Where(technicianPhoto => technicianPhoto.TechnicianId == technicianId)
+                .ToListAsync();
+        }
+
         public async Task<Technician> GetTechnician(int id)
         {
             return await _context.Technicians.FindAsync(id);
         }
-        
+
         public async Task Add(Technician technician)
         {
             await _context.Technicians.AddAsync(technician);
